requirements.txt
future
numpy
scipy
matplotlib
seaborn
onnx
netron
easydict
tensorboard
tqdm
pyyaml
opencv-python